+++ azext_iot/tests/test_constants.py
